moved setWindowsCommandAttr into process_windows.go
